requirements.txt
absl-py==0.8.0
astor==0.8.0
awsebcli==3.15.3
botocore==1.12.220
cement==2.8.2
certifi==2019.3.9
chardet==3.0.4
Click==7.0
colorama==0.3.9
docutils==0.15.2
Flask==1.0.2
future==0.16.0
gast==0.2.2
google-pasta==0.1.7
grpcio==1.23.0
h5py==2.9.0
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
Keras==2.2.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
numpy==1.17.1
pathspec==0.5.9
Pillow==6.1.0
protobuf==3.9.1
pypiwin32==223
python-dateutil==2.8.0
pywin32==224
PyYAML==3.13
requests==2.20.1
scipy==1.3.1
semantic-version==2.5.0
six==1.11.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
urllib3==1.24.3
virtualenv==16.4.0
wcwidth==0.1.7
Werkzeug==0.15.2
wincertstore==0.2
wrapt==1.11.2

